requirements.txt
streamlit==1.35.0
requests==2.31.0
google-auth==2.29.0
python-dotenv==1.0.1